# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Noroff EP e-commerce",
    "description": "APIs and a database to showcase their products for customers to purchase. Admin front-end, which uses the back-end system to manage the data."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "summary": "Fetch all brands",
        "description": "Fetches all brands from the database",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BrandsFetched"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "summary": "Create a new brand",
        "description": "Create a new brand. Requires admin authentication",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Brand name to be created",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateBrand"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/BrandCreated"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/brands/{id}": {
      "put": {
        "tags": [
          "Brands"
        ],
        "summary": "Update a brand",
        "description": "Update a brand name by ID. Requires admin authentication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the brand to update"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated brand name",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateBrand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BrandUpdated"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Brands"
        ],
        "summary": "Delete a brand",
        "description": "Delete a brand by ID. Requires admin authentication",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BrandDeleted"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/cart/": {
      "get": {
        "tags": [
          "Carts"
        ],
        "summary": "Fetch the users cart",
        "description": "Fetch logged-in users cart",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetCart"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Carts"
        ],
        "summary": "Add items to cart",
        "description": "Allows a logged-in user to add items to their cart",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Products to add to cart",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddToCart"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CartSuccess"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Carts"
        ],
        "summary": "Remove items from the cart or clear it",
        "description": "Removes a specific product from the cart if productId and quantity is provided, otherwise clear the cart",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "productId": {
                  "example": "any"
                },
                "quantity": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ProductRemoved"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/cart/checkout/now": {
      "post": {
        "tags": [
          "Carts"
        ],
        "summary": "Checkout and create order",
        "description": "Creates and checks out an order for logged-in user",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CheckoutOrder"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "BrandsFetched": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Viewing all brands"
            },
            "brands": {
              "type": "array",
              "example": [
                "List of all brands"
              ],
              "items": {
                "type": "string"
              }
            }
          },
          "required": [
            "result",
            "brands"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "CreateBrand": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Microsoft"
        }
      },
      "required": [
        "name"
      ]
    },
    "BrandCreated": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Created new brand"
            },
            "brand": {
              "type": "string",
              "example": "Microsoft"
            }
          },
          "required": [
            "result",
            "brand"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "UpdateBrand": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Amazon"
        }
      },
      "required": [
        "name"
      ]
    },
    "BrandUpdated": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Brand is updated"
            },
            "brand": {
              "type": "string",
              "example": "Amazon"
            }
          },
          "required": [
            "result",
            "brand"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "BrandDeleted": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "array",
              "example": [
                "Brand is deleted"
              ],
              "items": {
                "type": "string"
              }
            }
          },
          "required": [
            "result"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
...
    "GetCart": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Products in cart:"
            },
            "cart": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "ProductId": {
                    "type": "number",
                    "example": 8
                  },
                  "ProductName": {
                    "type": "string",
                    "example": "Samsung S20"
                  },
                  "Price": {
                    "type": "string",
                    "example": "900"
                  },
                  "Quantity": {
                    "type": "number",
                    "example": 2
                  },
                  "TotalPrice": {
                    "type": "number",
                    "example": 1800
                  }
                },
                "required": [
                  "ProductId",
                  "ProductName",
                  "Price",
                  "Quantity",
                  "TotalPrice"
                ]
              }
            }
          },
          "required": [
            "result",
            "cart"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "AddToCart": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "number",
          "example": 1
        },
        "quantity": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "productId",
        "quantity"
      ]
    },
    "CartSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Product added to cart"
            },
            "cart": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 10
                  },
                  "CartId": {
                    "type": "number",
                    "example": 1
                  },
                  "ProductId": {
                    "type": "number",
                    "example": 8
                  },
                  "Quantity": {
                    "type": "number",
                    "example": 2
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "Cart last updated"
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "Cart created"
                  }
                },
                "required": [
                  "id",
                  "CartId",
                  "ProductId",
                  "Quantity",
                  "updatedAt",
                  "createdAt"
                ]
              }
            }
          },
          "required": [
            "result",
            "cart"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
    "CheckoutOrder": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Order created."
            },
            "cart": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 5
                  },
                  "UserId": {
                    "type": "number",
                    "example": 2
                  },
                  "OrderNumber": {
                    "type": "string",
                    "example": "j8a72g31"
                  },
                  "Status": {
                    "type": "string",
                    "example": "In Progress"
                  },
                  "MembershipStatus": {
                    "type": "string",
                    "example": "Bronze"
                  },
                  "TotalPrice": {
                    "type": "string",
                    "example": "1800"
                  },
                  "updatedAt": {
                    "type": "string",
                    "example": "Order update"
                  },
                  "createdAt": {
                    "type": "string",
                    "example": "Ordered created"
                  }
                },
                "required": [
                  "id",
                  "UserId",
                  "OrderNumber",
                  "Status",
                  "MembershipStatus",
                  "TotalPrice",
                  "updatedAt",
                  "createdAt"
                ]
              }
            }
          },
          "required": [
            "result",
            "cart"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
...
    "ProductRemoved": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "success"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "data": {
          "type": "object",
          "properties": {
            "result": {
              "type": "string",
              "example": "Product removed from cart."
            }
          },
          "required": [
            "result"
          ]
        }
      },
      "required": [
        "status",
        "statusCode",
        "data"
      ]
    },
...
  }
}